# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/apps": {
            "get": {
                "description": "查询租户列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户接口"
                ],
                "summary": "查询租户列表接口",
                "operationId": "/apps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page index",
                        "name": "page_index",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ListAppsRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "创建租户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户接口"
                ],
                "summary": "创建租户接口",
                "operationId": "/apps",
                "parameters": [
                    {
                        "description": "create app request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrUpdateAppReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/po.App"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/apps/{app_id}": {
            "get": {
                "description": "查询租户详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户接口"
                ],
                "summary": "查询租户详情接口",
                "operationId": "/apps/{app_id}",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/po.App"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新租户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户接口"
                ],
                "summary": "更新租户接口",
                "operationId": "/apps/{app_id}",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update app request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrUpdateAppReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/po.App"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除租户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户接口"
                ],
                "summary": "删除租户接口",
                "operationId": "/apps/{app_id}",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/flow/apps/{app_id}": {
            "get": {
                "description": "查询租户流量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据接口"
                ],
                "summary": "查询租户流量接口",
                "operationId": "/flow/apps/{app_id}",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app id",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.Flow"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common_middleware.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        },
...
        "dto.CreateOrUpdateAppReq": {
            "type": "object",
            "properties": {
                "app_id": {
                    "description": "租户id",
                    "type": "string"
                },
                "app_name": {
                    "description": "租户名称",
                    "type": "string"
                },
                "qpd": {
                    "description": "每日请求量限制",
                    "type": "integer"
                },
                "qps": {
                    "description": "每秒请求量限制",
                    "type": "integer"
                },
                "secret": {
                    "description": "租户密钥",
                    "type": "string"
                },
                "white_ips": {
                    "description": "ip白名单，支持前缀匹配",
                    "type": "string"
                }
            }
        },
...
        "dto.Flow": {
            "type": "object",
            "properties": {
                "today_flow": {
                    "description": "今日流量列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "yesterday_flow": {
                    "description": "昨日流量列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "dto.ListAppItem": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "real_qpd": {
                    "type": "integer"
                },
                "real_qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
        "dto.ListAppsRes": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "租户列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ListAppItem"
                    }
                },
                "total": {
                    "description": "共计条数",
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "po.App": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "app_name": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
...
    }
}